movies/requirements.txt
opencv-python>=4.8.0
numpy>=1.21.0
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-multipart>=0.0.5
gradio>=4.0.0
ultralytics>=8.0.0
torch>=1.13.0
torchvision>=0.14.0
Pillow>=9.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
httpx>=0.24.0
websockets>=11.0
